fix #188, NODE_BC is a gateway message + response for Node id request, also add fix to ignore empty UID tag on node deletion

+++ modules/core/src/main/java/org/mycontroller/standalone/db/dao/UidTagDaoImpl.java
@@ -57,7 +57,9 @@ public class UidTagDaoImpl extends BaseAbstractDaoImpl<UidTag, Integer> implemen
     @Override
     public void deleteBySensorId(Integer sId) {
         List<Integer> senosrVariableIds = DaoUtils.getSensorVariableDao().getSensorVariableIds(sId);
-        deleteBySensorVariableIds(senosrVariableIds);
+        if (senosrVariableIds != null && !senosrVariableIds.isEmpty()) {
+            deleteBySensorVariableIds(senosrVariableIds);
+        }
     }
 
     @Override
